requirements.txt
beautifulsoup4==4.12.3
python_dateutil==2.9.0.post0
PyYAML==6.0.1
Requests==2.32.3
requests_cache==1.2.0
SQLAlchemy==2.0.30
psycopg2==2.9.9
